:wrench: Make DSAPI, concordances and Kafka all optional in config.
By default these are configured as enabled (meaning that the features
are configured in the service and included in the health check), but
they can be disabled individually (in which case the feature(s) are not
configured in the service and are excluded from the health check). If
all features have been disabled, a dummy health check is configured to
prevent warning messages on startup.