:art: :memo: Check for mode support; update README.
- If the document store or concordance API isn't enabled, then only
accept requests for suggest mode transformations, and gracefully reject
other modes.
- The UnsupportedTransformationModeException is reported to HTTP clients
with status 421 (Misdirected Request). Technically this is an HTTP/2
status code, but it seems the most appropriate (a request for preview or
publication should not have been made to a service thus configured).
- Updated the README to explain the transformation modes; also picked up
a few other things that were out of date.