:whale: Arbitrarily take one tag if there are many.
In general we don't expect multiple tags to point at the same commit,
but if, for example, we have to tag with different naming conventions to
deploy to different environments, then this may happen for legitimate
reason. Just take the first line of the git tag output, so that the
build works. In consequence the build-info may not contain the expected
tag, but it should contain an equivalent tag.