:bug: Allow configuration of source of audit field values.
In standard (UPP) environments, the audit field values (publishReference
and lastModified) should come from the current transaction. This is the
default configuration.

In PAC environments, the mapped content should contain values based on
the timestamp and reference of the native content rather than the
current transaction. This behaviour can be configured with Dropwizard
properties as described in the README.