:bug: Ignore unsupported properties regardless of type.
When an EomFile is received from Methode API, it has a property
`linkedObjects` that is an array. This property is not present at all
when EomFiles are retrieved from the native store. Prior to
63122622f071b3951272f44fbaf5899e59374dd0 this was handled by the
`@JsonIgnoreProperties` annotation, however if this is replaced by a
`@JsonAnySetter` then that setter method must handle values of any type.